voice_router/requirements.txt
fastapi==0.121.0
uvicorn-worker
python-multipart
openai-whisper
setuptools-rust
httpx
pydantic_settings
elasticsearch[async]==8.13.2
groq